Test data for the database that allows a login
The user is 1234567 and the course is 1.  There is a minor issue with the AttendanceRecords table.  There is a mistmatch in columns between the insert from Attendance/LoginAttendance.php and the columns in test.php.  test.php has the table AttendanceRecords with columns attendee_id ( no problem here) and attendance_key ( a mismatch).  A test shows no data being pumped into table AttendanceRecords.